move oracle logic into diff file
